homeTest/requirements.txt
nats_py==2.2.0
pandas==1.5.3
